Extract shared SubscriptionPeriodLabel component and fix locale bug
